+++ ChangeLog.md
@@ -2,7 +2,7 @@
 
 ### 3.3.0
 - ADDED: #3409 The Dark theme.
-- ADDED: #3471 Generated Columns support added.
+- ADDED: #3763 Generated Columns support added.
 - CHANGE: Windows binaries are now 64-bit.
 - CHANGE: SQLite2 plugin removed due to harder and harder maintenance of SQLite2 library compilation.
 - CHANGE: Minumum Qt version now is 5.12.0.